fixed more images improved editor-ui image parser

+++ editor-ui/src/Local-Site.coffee
@@ -28,7 +28,7 @@ class Local_Site
 
       files =
         for file in @.leanpub_Api.folder_content.files_Recursive()
-          if file.file_Extension() isnt '.jpg'
+          if file.file_Extension() isnt '.jpg' and file.file_Extension() isnt '.png'
             file.remove(@.leanpub_Api.folder_content.real_Path())
 
       if file_Name.file_Extension() is '.txt'
